Feat: 세션의 유저 아이디로 일정 수정 및 삭제 진행하도록 구현
- 로그인된 유저와 글 작성자가 다를 경우 throw

+++ src/main/java/com/example/scheduleprojectdevelop/service/ScheduleService.java
@@ -6,8 +6,10 @@ import com.example.scheduleprojectdevelop.entity.User;
 import com.example.scheduleprojectdevelop.repository.ScheduleRepository;
 import com.example.scheduleprojectdevelop.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
 
@@ -48,15 +50,25 @@ public class ScheduleService {
 
     //일정 수정
     @Transactional
-    public ScheduleResponseDto updateSchedule(Long id, String title, String contents) {
+    public ScheduleResponseDto updateSchedule(Long id, String title, String contents, Long userId) {
         Schedule schedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
-        schedule.update(title, contents);
-        return new ScheduleResponseDto(id, schedule.getTitle(), schedule.getContents());    }
+        //글 작성자와 현재 로그인된 유저 아이디가 같으면 수정
+        if((schedule.getUser().getId()).equals(userId)){
+            schedule.update(title, contents);
+            return new ScheduleResponseDto(id, schedule.getTitle(), schedule.getContents());
+        }
+        throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인의 글만 수정 가능합니다.");
+    }
 
 
     //일정 삭제
-    public void deleteSchedule(Long id) {
+    public void deleteSchedule(Long id, Long userId) {
         Schedule findSchedule = scheduleRepository.findScheduleByIdOrElseThrow(id);
-        scheduleRepository.delete(findSchedule);
+        //글 작성자와 현재 로그인된 유저의 아이디가 같으면 삭제
+        if((findSchedule.getUser().getId()).equals(userId)){
+            scheduleRepository.delete(findSchedule);
+        } else{
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "본인의 글만 삭제 가능합니다.");
+        }
     }
 }
